Move to mixin approach.
