+++ app/mixins/link-action.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-link-action/mixins/link-action';
